+++ SpringMicroservices/universitySE/src/main/java/com/example/universitySE/intservices/ProfessorServiceInterface.java
@@ -9,7 +9,7 @@ import java.util.List;
 
 public interface ProfessorServiceInterface {
 
-    SubjectModel getSubject(int id) throws SubjectException;
+    SubjectModel getSubject(long id) throws SubjectException;
 
     List<SupportMaterial> getMaterial() throws MaterialException;
