+++ SpringMicroservices/universitySE/src/main/java/com/example/universitySE/repositories/SecretaryRepository.java
@@ -1,7 +1,6 @@
 package com.example.universitySE.repositories;
 
 import com.example.universitySE.domain.*;
-import com.example.universitySE.dtos.ExamDTO;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -10,7 +9,7 @@ import java.util.Optional;
 @Repository
 public interface SecretaryRepository extends JpaRepository<Secretary, Integer> {
 
-    Optional<Secretary> findSecretaryById(int id);
+    Optional<Secretary> findSecretaryById(long id);
 
     void save(Person person);
