+++ SpringMicroservices/universitySE/src/main/java/com/example/universitySE/repositories/StudentRepository.java
@@ -9,6 +9,6 @@ import java.util.Optional;
 @Repository
 public interface StudentRepository extends JpaRepository<Student, Integer> {
 
-    Optional<Student> findStudentById(int id);
+    Optional<Student> findStudentById(long id);
 
 }
